requirements.txt
# requirements.txt for Python 3.8
Flask==2.2.5
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==1.4.46
Flask-Migrate==4.0.4
sqlalchemy-serializer==1.4.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
itsdangerous==2.1.2
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
gunicorn==21.0.0
requests==2.31.0
